fix(chat): sanitize persisted tool failures
