+++ apps/chat/src/app/api/chatbots/[chatbotId]/chat/route.ts
@@ -14,6 +14,10 @@ import {
   isAiTelemetryEnabled,
 } from '@/src/lib/server/langfuseTracing'
 import { getOpenAIResponsesStore } from '@/src/lib/server/openaiResponsesOptions'
+import {
+  mapAssistantStepContent,
+  type PersistedAssistantContentPart,
+} from '@/src/lib/server/persistedAssistantContent'
 import { CreditsService } from '@/src/services/credits'
 import { DisclaimersService } from '@/src/services/disclaimers'
 import {
@@ -569,133 +573,6 @@ const joinReasoningFromSteps = (
         : []
     )
     .join('\n\n')
-
-type PersistedAssistantContentPart =
-  | { type: 'text'; text: string }
-  | { type: 'reasoning'; text: string }
-  | {
-      type: 'tool-call'
-      toolCallId: string
-      toolName: string
-      args?: unknown
-      result?: unknown
-      isError?: boolean
-    }
-
-const mapAssistantStepContent = (
-  steps: Array<{ content?: unknown[] }> | undefined
-): PersistedAssistantContentPart[] => {
-  const content: PersistedAssistantContentPart[] = []
-  const toolCallIndexById = new Map<string, number>()
-
-  for (const step of steps ?? []) {
-    if (!Array.isArray(step.content)) continue
-
-    for (const rawPart of step.content) {
-      if (!rawPart || typeof rawPart !== 'object') continue
-
-      const part = rawPart as {
-        type?: unknown
-        text?: unknown
-        toolCallId?: unknown
-        toolName?: unknown
-        input?: unknown
-        output?: unknown
-        error?: unknown
-      }
-
-      if (part.type === 'text' && typeof part.text === 'string') {
-        content.push({ type: 'text', text: part.text })
-        continue
-      }
-
-      if (part.type === 'reasoning' && typeof part.text === 'string') {
-        content.push({ type: 'reasoning', text: part.text })
-        continue
-      }
-
-      if (
-        part.type === 'tool-call' &&
-        typeof part.toolCallId === 'string' &&
-        typeof part.toolName === 'string'
-      ) {
-        const nextToolCall = {
-          type: 'tool-call' as const,
-          toolCallId: part.toolCallId,
-          toolName: part.toolName,
-          args: part.input,
-        }
-        content.push(nextToolCall)
-        toolCallIndexById.set(nextToolCall.toolCallId, content.length - 1)
-        continue
-      }
-
-      if (part.type === 'tool-error' && typeof part.toolCallId === 'string') {
-        // A thrown tool execution error: without this branch the failure is
-        // silently dropped from the persisted message and reloads show a
-        // forever-loading tool call instead of the failed chip.
-        const errorText =
-          part.error instanceof Error
-            ? part.error.message
-            : typeof part.error === 'string'
-              ? part.error
-              : 'Tool execution failed'
-
-        const toolCallIndex = toolCallIndexById.get(part.toolCallId)
-        if (toolCallIndex !== undefined) {
-          const existingToolCall = content[toolCallIndex]
-          if (existingToolCall?.type === 'tool-call') {
-            existingToolCall.result = errorText
-            existingToolCall.isError = true
-          }
-          continue
-        }
-
-        if (typeof part.toolName !== 'string') {
-          continue
-        }
-
-        content.push({
-          type: 'tool-call',
-          toolCallId: part.toolCallId,
-          toolName: part.toolName,
-          args: part.input ?? {},
-          result: errorText,
-          isError: true,
-        })
-        toolCallIndexById.set(part.toolCallId, content.length - 1)
-        continue
-      }
-
-      if (part.type === 'tool-result' && typeof part.toolCallId === 'string') {
-        const toolCallIndex = toolCallIndexById.get(part.toolCallId)
-        if (toolCallIndex !== undefined) {
-          const existingToolCall = content[toolCallIndex]
-          if (existingToolCall?.type === 'tool-call') {
-            existingToolCall.result = part.output
-          }
-          continue
-        }
-
-        if (typeof part.toolName !== 'string') {
-          continue
-        }
-
-        const toolCallWithResult = {
-          type: 'tool-call' as const,
-          toolCallId: part.toolCallId,
-          toolName: part.toolName,
-          args: {},
-          result: part.output,
-        }
-        content.push(toolCallWithResult)
-        toolCallIndexById.set(part.toolCallId, content.length - 1)
-      }
-    }
-  }
-
-  return content
-}
 
 /**
  * Main chat endpoint that processes AI conversations with streaming responses.
